Modiy index.html and styles.css
I modified the index. html by setting up a page layout structure with for sections and a footer which include: header, and  main. The main section is diveided into three sections with a class of- section-1, section-2, and section-3. In css, I used a univer selector to reset padding and margin, and set box sizingto border box.To the body i set a font Roboto, display flex and flex direction column to send the footer to the bottom